hello

+++ src/pf_convertions/x_upper.c
@@ -6,7 +6,7 @@
 /*   By: jandre-d <jandre-d@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/04/17 20:47:33 by jandre-d       #+#    #+#                */
-/*   Updated: 2019/04/22 18:10:40 by jandre-d      ########   odam.nl         */
+/*   Updated: 2019/04/25 14:40:34 by jandre-d      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,6 @@
 
 bool	pf_x_upper(t_conversion_in *c_in, t_conversion_out *c_out, va_list *argl)
 {
-	c_out->str = pf_uitoa_base(va_arg(*argl, unsigned int), 16, &c_out->len, false);
+	c_out->str = pf_uitoa_base(va_arg(*argl, uint64_t), 16, &c_out->len, false);
 	return (c_out->str != NULL);
 }
